// types.ts
export enum CommandType {
  // Media Controls
  MEDIA_PLAY_PAUSE = 'MEDIA_PLAY_PAUSE',
  MEDIA_NEXT = 'MEDIA_NEXT',
  MEDIA_PREVIOUS = 'MEDIA_PREVIOUS',
  MEDIA_VOLUME_UP = 'MEDIA_VOLUME_UP',
  MEDIA_VOLUME_DOWN = 'MEDIA_VOLUME_DOWN',
  MEDIA_MUTE = 'MEDIA_MUTE',

  // Power Controls
  POWER_SHUTDOWN = 'POWER_SHUTDOWN',
  POWER_RESTART = 'POWER_RESTART',
  POWER_SLEEP = 'POWER_SLEEP',

  // Keyboard Control
  KEYBOARD_TYPE = 'KEYBOARD_TYPE',
  
  // Shortcut Controls
  OPEN_BROWSER = 'OPEN_BROWSER',
  OPEN_TERMINAL = 'OPEN_TERMINAL',
}